+++ client/src/pages/Top/Content.tsx
@@ -2,17 +2,14 @@ import React from 'react';
 import styled from 'styled-components/macro';
 import Grid from '@material-ui/core/Grid';
 import Avatar from '@material-ui/core/Avatar';
-import Typography from '@material-ui/core/Typography';
+import ContentFooter from './ContentFooter';
+import ContentHeader from './ContentHeader';
 
 const Content = styled.div`
   border-bottom: solid 1px;
   border-color: rgb(61, 84, 102);
   flex-grow: 1;
   padding: 0 10px 0 10px;
-`;
-
-const GrayTypography = styled(Typography)`
-  color: gray;
 `;
 
 export default () => {
@@ -23,15 +20,21 @@ export default () => {
           <Avatar src="/images/blue-bird.png" alt="青い鳥" />
         </Grid>
         <Grid item container>
-          <Grid item xs={4} zeroMinWidth>
-            <Typography noWrap>hogehogehogehoge</Typography>
+          <ContentHeader />
+          <Grid item container>
+            槍が岳に行ってきました
+            <br />
+            最高だぜ！！！
+            <br />
           </Grid>
-          <Grid item xs={4} zeroMinWidth>
-            <GrayTypography noWrap>@hogehoge</GrayTypography>
+          <Grid item container>
+            <img
+              src="https://www.hamchance.com/images/yarigadake.jpg"
+              alt="槍ヶ岳"
+              width="100%"
+            />
           </Grid>
-          <Grid item xs={4}>
-            <GrayTypography>1時間前</GrayTypography>
-          </Grid>
+          <ContentFooter />
         </Grid>
       </Grid>
     </Content>
