+++ client/src/pages/Top/Contents.tsx
@@ -17,84 +17,6 @@ export default () => {
       <Content />
       <Content />
       <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
-      <Content />
     </Contents>
   );
 };
